requirements.txt
tensorflow>=2.15.0
fastapi
uvicorn
streamlit
requests
opencv-python-headless
pillow
numpy